More refactoring InventoryList
