+++ EventsExpress/ClientApp/src/actions/inventory-list.js
@@ -11,6 +11,7 @@ export function get_inventories_by_event_id(eventId) {
         dispatch(setInventoryPending(true));
         const res = api_serv.getInventoriesByEventId(eventId);
         res.then(response => {
+            console.log('action', response);
             if (response.error == null) {
                 dispatch(getInventorySuccess(response));
             } else {
