+++ EventsExpress/ClientApp/src/components/inventory/InventoryList.js
@@ -1,15 +1,13 @@
 import React, { Component } from 'react';
 import InventoryHeaderButton from './InventoryHeaderButton';
-import OwnerEditItemForm from './ownerEditItem';
-import VisitorEditItemForm from './visitorTakeItem';
 import { connect } from 'react-redux';
 import  get_unitsOfMeasuring  from '../../actions/unitsOfMeasuring';
-import { update_inventories, get_inventories_by_event_id }  from '../../actions/inventory-list';
-import { get_users_inventories_by_event_id, delete_users_inventory, edit_users_inventory }  from '../../actions/usersInventories';
-import { add_item, delete_item, edit_item, want_to_take } from '../../actions/inventar';
 import IconButton from "@material-ui/core/IconButton";
-import OwnerSeeItem from './ownerSeeItem';
-import VisitorSeeItem from './VisitorSeeItem';
+import { get_inventories_by_event_id }  from '../../actions/inventory-list';
+import InventoryItemWrapper from '../../containers/inventory-item';
+import { update_inventories }  from '../../actions/inventory-list';
+import { get_users_inventories_by_event_id, edit_users_inventory }  from '../../actions/usersInventories';
+import { add_item, edit_item } from '../../actions/inventar';
 
 class InventoryList extends Component {
 
@@ -21,12 +19,8 @@ class InventoryList extends Component {
             showAlreadyGetDetailed: false
         };
 
-        this.handleOnClickCaret = this.handleOnClickCaret.bind(this);
-        this.handleOnClickWantToTake = this.handleOnClickWantToTake.bind(this);
-        this.onAlreadyGet = this.onAlreadyGet.bind(this);
         this.onSubmit = this.onSubmit.bind(this);
-        this.onCancel = this.onCancel.bind(this);
-        this.onWillTake = this.onWillTake.bind(this);
+        this.handleOnClickCaret = this.handleOnClickCaret.bind(this);
     }
 
     componentDidMount() {
@@ -51,54 +45,21 @@ class InventoryList extends Component {
         });
     }
 
-    deleteItemFromList = inventar => {
-        this.props.delete_item(inventar.id, this.props.eventId);
-    }
-
-    markItemAsEdit = inventar => {
-        let updateList = this.props.inventories.items;
-        updateList.find(e => e.id === inventar.id).isEdit = true;
-
-        this.props.get_inventories(updateList);
-        this.setState({
-            disabledEdit: true
-        });
-    }
-
-    markItemAsWillTake = inventar => {
-        let updateList = this.props.inventories.items;
-        updateList.find(e => e.id === inventar.id).isWillTake = true;
-        updateList.find(e => e.id === inventar.id).isEdit = true;
-
-        this.props.get_inventories(updateList);
-        this.setState({
-            disabledEdit: true
-        });
-    }
-
-    onAlreadyGet = inventar => {
-        let updateList = this.props.inventories.items;
-        updateList.map(item => {
-            if (inventar.id === item.id)
-                item.showAlreadyGetDetailed = !item.showAlreadyGetDetailed;
-        });
-
-        this.props.get_inventories(updateList);
-    }
-
     handleOnClickCaret = () => {
         this.setState(state => ({
             isOpen: !state.isOpen
         }));
     }
 
-    handleOnClickWantToTake = () => {
-        this.setState(state => ({
-            showAlreadyGetDetailed: !state.showAlreadyGetDetailed
-        }));
+    changeDisableEdit = (value) => {
+        this.setState({
+            disabledEdit: value
+        });
     }
 
     onSubmit = values => {
+        console.log('submit', values);
+
         if (values.isNew) {
             this.props.add_item(values, this.props.eventId);
         }
@@ -112,103 +73,6 @@ class InventoryList extends Component {
         });
     }
 
-    onCancel = inventar => {
-        if (inventar.isNew) {
-            this.deleteItemFromList(inventar);
-            this.setState({
-                disabledEdit: false
-            })
-            return;
-        }
-
-        this.props.get_inventories_by_event_id(this.props.eventId);
-
-        this.setState({
-            disabledEdit: false
-        });
-    }
-
-    onWillTake = inventar => {
-        const data = {
-            eventId: this.props.eventId,
-            userId: this.props.user.id,
-            inventoryId: inventar.id,
-            quantity: Number(inventar.willTake)
-        }
-
-        if (inventar.isWillTake) {
-            this.props.want_to_take(data);
-        } else {
-            this.props.edit_users_inventory(data);
-        }
-
-        this.setState({
-            disabledEdit: false
-        });
-    }
-
-    onWillNotTake = inventar => {
-        const data = {
-            eventId: this.props.eventId,
-            userId: this.props.user.id,
-            inventoryId: inventar.id
-        }
-
-        this.props.delete_users_inventory(data);
-    }
-
-    renderInventoryItem = item => {
-        const { user, usersInventories } = this.props;
-        let isMyEvent = this.props.event.owners.find(x => x.id === user.id) != undefined;
-        return (
-            <div className="row p-1 d-flex align-items-center" key={item.id}>
-                {item.isEdit && isMyEvent && 
-                    <OwnerEditItemForm
-                        onSubmit={this.onSubmit} 
-                        onCancel={this.onCancel}
-                        unitOfMeasuringState={this.props.unitOfMeasuringState}
-                        alreadyGet={usersInventories.data.reduce((acc, cur) => {
-                            return cur.inventoryId === item.id ? acc + cur.quantity : acc + 0
-                        }, 0)}
-                        initialValues={item}/>
-                }
-
-                {item.isEdit && !isMyEvent &&
-                    <VisitorEditItemForm
-                        onSubmit={this.onWillTake}
-                        onCancel={this.onCancel}
-                        alreadyGet={usersInventories.data.reduce((acc, cur) => {
-                            return cur.inventoryId === item.id ? acc + cur.quantity : acc + 0
-                        }, 0) - (item.isWillTake ? 0 : usersInventories.data.find(e => e.inventoryId === item.id)?.quantity || 0)}
-                        initialValues={item}
-                    />
-                }
-
-                {!item.isEdit && isMyEvent &&
-                    <OwnerSeeItem
-                        item={item}
-                        disabledEdit={this.state.disabledEdit}
-                        onAlreadyGet={this.onAlreadyGet}
-                        markItemAsEdit={this.markItemAsEdit}
-                        deleteItemFromList={this.deleteItemFromList}
-                        usersInventories={this.props.usersInventories}/>
-                }
-
-                {!item.isEdit && !isMyEvent && 
-                    <VisitorSeeItem 
-                        item={item}
-                        disabledEdit={this.state.disabledEdit}
-                        onWillNotTake={this.onWillNotTake}
-                        markItemAsEdit={this.markItemAsEdit}
-                        markItemAsWillTake={this.markItemAsWillTake}
-                        usersInventories={this.props.usersInventories}
-                        user={this.props.user}/>
-                }
-            </div>
-        )
-}
-
-
     render() {
         const { inventories, event, user, usersInventories } = this.props;
         let isMyEvent = event.owners.find(x => x.id === user.id) != undefined;
@@ -219,7 +83,7 @@ class InventoryList extends Component {
                     ...item,
                     isTaken: usersInventories.data
                         .filter(dataItem => 
-                            this.props.user.id === dataItem.userId && 
+                            user.id === dataItem.userId && 
                             item.id === dataItem.inventoryId)
                         .length > 0
                 }                
@@ -255,7 +119,24 @@ class InventoryList extends Component {
                                 <div className="col col-md-1"></div>
                                 <div className="col col-md-2"></div>
                             </div>
-                            {updateList.map(item => this.renderInventoryItem(item))}                    
+                            {updateList.map((item, key) => {
+                                return (
+                                    <InventoryItemWrapper
+                                        item={item}
+                                        user={user}
+                                        usersInventories={usersInventories}
+                                        inventories={inventories}
+                                        isMyEvent={isMyEvent}
+                                        disabledEdit={this.state.disabledEdit}
+                                        onSubmit={this.onSubmit}
+                                        changeDisableEdit={this.changeDisableEdit}
+                                        get_inventories={this.props.get_inventories}
+                                        event={event}
+                                        eventId={this.props.eventId}
+                                        key={key}
+                                    />
+                                );
+                            })}                    
                         </div>
                 </div>
                 }
@@ -265,7 +146,6 @@ class InventoryList extends Component {
 }
 
 const mapStateToProps = (state) => ({
-    unitOfMeasuringState: state.unitsOfMeasuring,
     event: state.event.data,
     user: state.user,
     inventories: state.inventories,
@@ -276,14 +156,11 @@ const mapDispatchToProps = (dispatch) => {
     return {
         get_unitsOfMeasuring: () => dispatch(get_unitsOfMeasuring()),
         add_item: (item, eventId) => dispatch(add_item(item, eventId)),
-        delete_item: (itemId, eventId) => dispatch(delete_item(itemId, eventId)),
         edit_item: (item, eventId) => dispatch(edit_item(item, eventId)),
         get_inventories: (inventories) => dispatch(update_inventories(inventories)),
-        get_inventories_by_event_id: (eventId) => dispatch(get_inventories_by_event_id(eventId)),
         get_users_inventories_by_event_id: (eventId) => dispatch(get_users_inventories_by_event_id(eventId)),
-        delete_users_inventory: (data) => dispatch(delete_users_inventory(data)),
-        edit_users_inventory: (data) => dispatch(edit_users_inventory(data)),
-        want_to_take: (data) => dispatch(want_to_take(data))
+        get_inventories_by_event_id: (eventId) => dispatch(get_inventories_by_event_id(eventId)),
+        edit_users_inventory: (data) => dispatch(edit_users_inventory(data))
     }
 };
