+++ EventsExpress/Mapping/AutoMapperProfile.cs
@@ -325,26 +325,6 @@ namespace EventsExpress.Mapping
             CreateMap<UnitOfMeasuring, UnitOfMeasuringDTO>().ReverseMap();
             CreateMap<UnitOfMeasuringDTO, UnitOfMeasuringViewModel>().ReverseMap();
             #endregion
-
-            #region USEREVENTINVENTORY MAPPING
-            CreateMap<UserEventInventory, UserEventInventoryDTO>()
-                .ForMember(dest => dest.User, opt => opt.MapFrom(src => new UserDTO
-                {
-                    Id = src.UserId,
-                    Name = src.UserEvent.User.Name,
-                }));
-            CreateMap<UserEventInventoryDTO, UserEventInventoryViewModel>()
-                .ForMember(dest => dest.User, opt => opt.MapFrom(src => new UserPreviewViewModel
-                {
-                    Id = src.User.Id,
-                    Username = src.User.Name,
-                }));
-            CreateMap<UserEventInventoryViewModel, UserEventInventoryDTO>()
-                .ForMember(dest => dest.User, opt => opt.Ignore());
-            CreateMap<UserEventInventoryDTO, UserEventInventory>()
-                .ForMember(dest => dest.UserEvent, opt => opt.Ignore())
-                .ForMember(dest => dest.Inventory, opt => opt.Ignore());
-            #endregion
         }
     }
 }
