+++ EventsExpress/Validation/InventoryViewModelValidator.cs
@@ -9,7 +9,6 @@ namespace EventsExpress.Validation
         {
             RuleFor(x => x.ItemName).NotEmpty().Length(1, 30);
             RuleFor(x => x.NeedQuantity).GreaterThan(0);
-            RuleFor(x => x.UnitOfMeasuring.Id).NotEmpty();
         }
     }
 }
